requirements.txt
requests==2.26.0
pytz==2021.3
python-dotenv==0.19.2
schedule==1.1.0